Exclude logs from otel and refactor into their own extensions
